refactor(button): consolidate on single Button component
Delete legacy button/index.tsx and button/button-simple.tsx; migrate all
remaining callers (webhooks create, newbie content, otp-box) to the
feature-complete button/button.tsx with the standard isLoading prop.
